Backup before rewriting activation method
Parameters for single FS were tuned to obtain activation only when input
string matches target string (see AtomFS_dyn.py). It is planned to hide
parameters of activation and prediction methods of AtomFS . Two kinds of
inputs to the atom FS should be implemented, these are perception
(problem) inputs that activate FS bottom up and control inputs for top
down activation. The former are normalized but the latter ones not.